Don't fail containerd start in case of corrupted events.log
Some host crash scenarious might cause a corruption at the end of the
events.log. In those cases, the file ends with multiple 0x00 (assuming
that were created for journaling of an event prior the crash).
In such case, there is no meaning of preventing containerd from starting.
This commit add a logic of ignoring the rest of the file (it should be a single event
that is missing anyhow) and proceed with those that were succesfully parsed so far.

Tested using the corrupted log attached to the github issue.

Signed-off-by: Rom Freiman <rom@stratoscale.com>
(cherry picked from commit 3b3a98cca1d0023a2005141948508cd610d72485)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>